Fix models and Db

+++ ReportingService.DB/dbo/Tables/LeadStatistic.sql
@@ -2,14 +2,14 @@
 (
 	Id bigint NOT NULL,
 	DateStatistic date NOT NULL,
-	LeadId bigint NOT NULL,
+	LeadId int NOT NULL,
 	TransactionCountForTwoMonth int NOT NULL,
 	DepositsSum decimal(14,4) NOT NULL,
 	WithdrawSum decimal(14,4) NOT NULL,
 	TransferSum decimal(14,4) NOT NULL,
-	DepositCount bigint NOT NULL,
-	WithdrawCount bigint NOT NULL,
-	TransferCount bigint NOT NULL,
+	DepositCount int NOT NULL,
+	WithdrawCount int NOT NULL,
+	TransferCount int NOT NULL,
   CONSTRAINT [PK_LEADSTATISTIC] PRIMARY KEY CLUSTERED
   (
   [Id] ASC
