+++ ReportingService.Data/Dto/LeadStatisticDto.cs
@@ -5,7 +5,7 @@ public class LeadStatisticDto
 {
     public long Id { get; set; }
     public DateOnly DateStatistic { get; set; }
-    public long LeadId { get; set; }
+    public int LeadId { get; set; }
     public int TransactionCountForTwoMontht {get; set; }
     public decimal DepositsSum { get; set; }
     public decimal WithdrawSum { get; set; }
